Refactoring the code into a DumpMeta class
