+++ dgraph/cmd/migrate/utils.go
@@ -17,9 +17,12 @@
 package migrate
 
 import (
+	"bufio"
 	"database/sql"
 	"fmt"
+	"os"
 	"reflect"
+	"sort"
 	"strings"
 
 	_ "github.com/go-sql-driver/mysql"
@@ -88,10 +91,6 @@ type ColumnIdx struct {
 	index int    // the column index
 }
 
-func getLinkPredicate(predicate string) string {
-	return "mysql." + predicate
-}
-
 // ptrToValues takes a slice of pointers, deference them, and return the values referenced by these
 // pointers
 func ptrToValues(ptrs []interface{}) []interface{} {
@@ -102,4 +101,46 @@ func ptrToValues(ptrs []interface{}) []interface{} {
 		values = append(values, v)
 	}
 	return values
+}
+
+func getFileWriter(filename string) (*bufio.Writer, func(), error) {
+	output, err := os.OpenFile(filename, os.O_WRONLY|os.O_CREATE|os.O_TRUNC, 0600)
+	if err != nil {
+		return nil, nil, err
+	}
+
+	return bufio.NewWriter(output), func() { output.Close() }, nil
+}
+
+func getColumnValues(columns []string, columnTypes []*sql.ColumnType, rows *sql.Rows) []interface{} {
+	colValuePtrs := make([]interface{}, 0, len(columns))
+	for i := 0; i < len(columns); i++ {
+		switch columnTypes[i].DatabaseTypeName() {
+		case "VARCHAR":
+			colValuePtrs = append(colValuePtrs, new([]byte)) // the value can be nil
+		case "INT":
+			colValuePtrs = append(colValuePtrs, new(int))
+		case "FLOAT":
+			colValuePtrs = append(colValuePtrs, new(float64))
+		default:
+			panic(fmt.Sprintf("unknown type %v at index %d",
+				columnTypes[i].ScanType().Kind(), i))
+		}
+	}
+	rows.Scan(colValuePtrs...)
+	colValues := ptrToValues(colValuePtrs)
+	return colValues
+}
+
+// getSortedColumns sorts the column alphabetically using the column names
+// and return the column names as a slice
+func getSortedColumns(tableInfo *TableInfo) []string {
+	columns := make([]string, 0)
+	for column := range tableInfo.columns {
+		columns = append(columns, column)
+	}
+	sort.Slice(columns, func(i, j int) bool {
+		return columns[i] < columns[j]
+	})
+	return columns
 }
